+++ docs/abstract-algebra/development/source/conf.py
@@ -14,8 +14,8 @@ project = 'Boring Math - Abstract Algebra'
 copyright = '2025, Geoffrey R. Scheller'
 author = 'Geoffrey R. Scheller'
 release = '0.1.0' 
-release_string = 'Proposed PyPI'
-# release_string = 'PyPI'
+# release_string = 'Proposed PyPI'
+release_string = 'PyPI'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -30,6 +30,9 @@ exclude_patterns: list[str] = []
 
 # -- Options for Sphinx
 autoclass_content = 'both'
+
+# -- Options for autodoc extension
+autodoc_member_order = 'bysource'
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output